single node process
