--- dashboard/config/dashboard.ini
@@ -1,8 +1,0 @@
-[web]
-port = 8081
-api  = http://35.167.50.63:8080/
-client = http://35.167.50.63/sugarizer/
-
-[activities]
-activities_directory_name = activities
-activities_path = ../../sugarizer/activities
