--- dashboard/package.json
@@ -1,18 +1,0 @@
-{
-	"name": "sugarizer-server-dashboard",
-	"description": "Sugarizer Server Dashboard",
-	"version": "0.0.1",
-	"dependencies": {
-		"express": "3.x",
-		"http": "0.0.x",
-		"ini": "1.1.x",
-    "ejs": "^1.0.0",
-    "moment": "2.x.x",
-    "express-flash": "~0.0.2",
-    "express-validator": "~1.0.1",
-    "request" : "*"
-	},
-	"scripts": {
-		"run": "node dashboard.js"
-	}
-}
